+++ src/NukedOpl/Opl3Writebuf.cs
@@ -42,7 +42,7 @@ namespace NukedOpl
     public sealed class Opl3Writebuf
     {
         public ulong time { get; set; }
-        public ushort reg { get; set; }
+        public int reg { get; set; }
         public byte data { get; set; }
 
         public void Reset()
